Fixed exception on Argument.TryParseToValue
GetFullyQualifiedName() doesn't work with ObjectCreationExpressionSyntax when it's part of another project.
As it was just to fill something in the 'Object' of the Value, I've changed it to ToFullString() instead.
Added extra test cases to test other projects. Changed the SolutionBuilderHelper to be able to create a solution with 2 projects.